vectorfit is facetable + fix exports + fix current bugs
The vectorfit faceting is still a crude hack, but it seems to work,
and I polish it later.

StatVectorfit must be exported or function fails. This worked in
my test, because I had a stray StatVectorfit on my work space,
but failed in clean environment.

Stupid editing bugs remained after splitting geom_ordi[text/label].